Starting to rank hands by Poker rules
(ab)Use our sorted list of cards and "count of suits" for Flushes.

Q. Since we have a static hand, would it not be better to put this
into the constructor rather than calculating it later?  Do we gain
anything by having this "on the fly"?  (A: If we have a river/flop,
probably, since then cards not in the hand can influence the rank,
maybe.  If I remember correctly.  Which I frequently don't with Poker.)