Final on-the-fly ranking code
This feels like terrible cheating and I'm sure there's a cleverer way
to do this - representing a hand as a collection of 5 6-bit values
(2 bits for hand, 4 bits for card) with the appropriate bitmasking and
comparisons might work.  Haven't really thought that through.  On the
other hand, this is a 2 hour project and getting something running is
vastly more important than making a lovely crystal space castle.

I hope.